chore(broker-core): move Environment and TomlConfigurationReader to util
- prepare configuration utils to be reusable from gateway
