+++ broker-core/src/main/java/io/zeebe/broker/system/configuration/DataCfg.java
@@ -17,6 +17,7 @@
  */
 package io.zeebe.broker.system.configuration;
 
+import io.zeebe.util.Environment;
 import java.util.Arrays;
 import java.util.List;
