+++ broker-core/src/main/java/io/zeebe/broker/system/configuration/ExporterCfg.java
@@ -17,6 +17,7 @@
  */
 package io.zeebe.broker.system.configuration;
 
+import io.zeebe.util.Environment;
 import java.util.Map;
 
 /**
